+++ src/commonMain/kotlin/kotools/types/experimental/Annotations.kt
@@ -89,9 +89,13 @@ public annotation class ExperimentalRangeApi
     ReplaceWith(
         RECOMMENDED_ANNOTATION,
         "$RECOMMENDED_ANNOTATION_PACKAGE.$RECOMMENDED_ANNOTATION"
-    )
+    ),
+    DeprecationLevel.ERROR
 )
-@DeprecatedSince(KotoolsTypesVersion.V4_4_0)
+@DeprecatedSince(
+    warningSince = KotoolsTypesVersion.V4_4_0,
+    errorSince = KotoolsTypesVersion.Unreleased
+)
 @MustBeDocumented
 @RequiresOptIn(OPT_IN_MESSAGE)
 @Retention(BINARY)