+++ src/commonTest/kotlin/org/kotools/types/web/EmailAddressTest.kt
@@ -1,4 +1,4 @@
-package kotools.types.experimental
+package org.kotools.types.web
 
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.SerializationException
@@ -8,6 +8,7 @@ import kotlinx.serialization.descriptors.SerialKind
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.serializer
+import kotools.types.experimental.ExperimentalKotoolsTypesApi
 import kotools.types.internal.ErrorMessage
 import kotools.types.internal.KotoolsTypesPackage
 import kotools.types.internal.deserializationErrorMessage