+++ subprojects/java-compatibility/src/test/java/org/kotools/types/web/EmailAddressTest.java
@@ -1,4 +1,4 @@
-package kotools.types.experimental;
+package org.kotools.types.web;
 
 import kotlin.text.Regex;
 import org.junit.jupiter.api.Assertions;