+++ gradle/plugins/src/main/kotlin/kotools/types/plugins/Multiplatform.kt
@@ -31,6 +31,7 @@ private fun PluginContainer.configureYarn(project: Project): Unit =
     withType<YarnPlugin>().configureEach {
         val yarn: YarnRootExtension = project.rootProject.extensions.getByType()
         yarn.lockFileDirectory = project.rootDir
+        yarn.resolution("follow-redirects", "1.15.4")
         yarn.resolution("webpack", "5.76.3")
     }