+++ subprojects/java-compatibility/src/test/java/kotools/types/web/EmailAddressTest.java
@@ -1,4 +1,4 @@
-package org.kotools.types.web;
+package kotools.types.web;
 
 import kotlin.text.Regex;
 import org.junit.jupiter.api.Assertions;