+++ CHANGELOG.md
@@ -28,8 +28,9 @@ Documentation of dependency compatibility (issue [#288]).
 
 ### Deprecated
 
-Deprecation promotion of the `ExperimentalCollectionApi` annotation to error
-(issue [#333](https://github.com/kotools/types/issues/333)).
+Deprecation promotion of the `ExperimentalCollectionApi` and the
+`ExperimentalNumberApi` annotations to error (issue
+[#333](https://github.com/kotools/types/issues/333)).
 
 ### Security