+++ CHANGELOG.md
@@ -32,7 +32,8 @@ Deprecation promotion of the following annotations to error (issue
 [#333](https://github.com/kotools/types/issues/333)):
 - `ExperimentalCollectionApi`
 - `ExperimentalNumberApi`
-- `ExperimentalRangeApi`.
+- `ExperimentalRangeApi`
+- `ExperimentalResultApi`
 
 ### Security