+++ CHANGELOG.md
@@ -32,8 +32,8 @@ Documentation of dependency compatibility (issue
 - Upgrade kotlinx.serialization to 1.5.0 (issue
   [#378](https://github.com/kotools/types/issues/378)).
 - Move the `EmailAddress` **experimental** type from the
-  `kotools.types.experimental` package to a new `org.kotools.types.web` one
-  (issue [#377](https://github.com/kotools/types/issues/377)).
+  `kotools.types.experimental` package to the new `kotools.types.web` one (issue
+  [#377](https://github.com/kotools/types/issues/377)).
 
 ### Deprecated