+++ CHANGELOG.md
@@ -62,6 +62,22 @@ Improper Input Validation due to the improper handling of URLs by the
 See the [security report](https://github.com/advisories/GHSA-jchw-25xp-jwwc) for
 more details on this vulnerability.
 
+## 4.4.1 <a id="4.4.1"></a>
+
+_Release date: 2024-02-02 | Commits: [4.4.0...4.4.1]._
+
+[4.4.0...4.4.1]: https://github.com/kotools/types/compare/4.4.0...4.4.1
+
+### Fixed
+
+This release fixes the publication of the previous one, in which the
+`types-internal` project was missing in the Gradle dependency tree.
+
+So it is recommended to bypass version 4.4.0 and directly upgrade to 4.4.1.
+
+Thanks to [@robertfmurdock](https://github.com/robertfmurdock) for reporting
+this error (issue [#406](https://github.com/kotools/types/issues/406)).
+
 ## 4.4.0 <a id="4.4.0"></a>
 
 _Release date: 2024-01-29 | Commits: [4.3.1...4.4.0]._